Moving the PlanMix in functionality from view to mixin. Version - 0.0.1
