+++ anime/api/views.py
@@ -6,7 +6,7 @@ from anime.api.filters import AnimeFilter
 from anime.api.serializers import AnimeListSerializer
 from anime.documents import AnimeDocument
 from anime.models import Anime
-from plans.api.views import PlanMixin
+from plans.api.mixin import PlanMixin
 from search.api.views import PaginatedElasticSearchAPIView
 from comments.api.views import CommentMixin
 from rating.api.views import RatingMixin
